use a dict instead of a class to store state
Currently the UI stores all its state in the global namespace. To make
this cleaner I would need to refactor all of the code into some classes,
but since this is a hobby project and I don't need to add anything else
to the UI for the moment i'm storing state in a dict.